@@ -13,36 +13,42 @@
     "test:ci": "jest -i --ci --runInBand"
   },
   "devDependencies": {
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
     "@testing-library/jest-dom": "^5.11.0",
     "@testing-library/react": "^10.4.3",
     "@testing-library/user-event": "^12.0.11",
     "@types/jest": "^26.0.3",
     "@types/node": "^8.0.14",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
+    "babel-loader": "^8.2.5",
     "eslint": "^7.4.0",
+    "html-webpack-plugin": "^5.5.0",
     "jest": "^26.1.0",
     "jest-css-modules": "^2.1.0",
     "jest-junit": "^11.0.1",
-    "sass": "^1.26.9",
-    "sass-loader": "^8.0.2",
+    "react-refresh": "^0.14.0",
     "ts-jest": "^26.1.1",
-    "typescript": "^3.9.6",
-    "webpack": "^4.43.0"
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
+    "@babel/plugin-syntax-jsx": "^7.18.6",
+    "@babel/preset-typescript": "^7.18.6",
     "@types/react-router-config": "^5.0.1",
     "css-loader": "^3.6.0",
-    "express": "^4.17.1",
-    "file-loader": "^6.0.0",
     "path": "^0.12.7",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-config": "^5.1.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.4.1",
+    "run": "^1.4.0",
+    "sass": "^1.26.9",
+    "sass-loader": "^8.0.2",
     "style-loader": "^1.2.1",
-    "ts-loader": "^7.0.5",
-    "webpack-cli": "^3.3.12"
+    "ts-loader": "^9.4.1",
+    "typescript": "^4.8.3",
+    "webpack": "^5.2.0",
+    "webpack-cli": "^4.10.0"
   },
   "license": "ISC",
   "optionalDependencies": {
