@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Square from './Square';
 import '../../stylesheet/components/Board.scss';
 
