@@ -1,8 +1,8 @@
-import * as React from 'react';
+import React from 'react';
 
 type HelloState = { count: number };
 
-export default class Hello extends React.Component<unknown, HelloState> {
+export class Hello extends React.Component<unknown, HelloState> {
     constructor(_: unknown) {
         super(_);
 
