@@ -1,8 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-
-import Hello from '../Hello';
+import { Hello } from '../Hello';
 
 // サンプル用テストコード
 
@@ -16,7 +15,7 @@ describe('HelloTest', () => {
 
     test('count should increase by 1 when button is clicked', () => {
         render(<Hello />);
-        
+
         userEvent.click(screen.getByRole('button'));
         expect(screen.getByText(`count: ${1}`)).toBeInTheDocument();
 
